+++ CHANGELOG.md
@@ -2,9 +2,41 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-## [0.1.0](https://github.com/Amsterdam/Vergunningschecker/compare/v0.0.8...v0.1.0) (2020-01-09)
-
+### [0.1.1](https://github.com/Amsterdam/Vergunningschecker/compare/v0.1.0...v0.1.1) (2020-01-15)
 
 ### Features
 
-* Added more standard version ([1e9f67e](https://github.com/Amsterdam/Vergunningschecker/commit/1e9f67eb1fc89f8082dd2428c92077df24fd3886))
+- LocationPage styling and text
+- Header styling and text
+- Footer styling and text
+- NotFoundPage styling and text
+- Added `nextText` prop to `Navigation`
+- Improved redirects to `oloChecker`
+- Added functions to `GET_CURRENT_TOPIC` and `GET_CURRENT_PAGE`
+- Added function to `GET_TEXT` based on `Topic`
+- Updated Routes and Links
+- Fixed Bestemmingsplan bugs
+- Fix Momument fetch #124
+- Fix LocationPage input bugs
+
+### Bug Fixes
+
+- 19c now works :-) ([cc51938](https://github.com/Amsterdam/Vergunningschecker/commit/cc51938e92950c6e3730b6c7777bf0052583c6b4))
+- also metrostation weesperplein works ([47b81f7](https://github.com/Amsterdam/Vergunningschecker/commit/47b81f769c15e98d8628d54907fcabef08f4a3ae))
+- Bug in housenummer suffix ([8c949a6](https://github.com/Amsterdam/Vergunningschecker/commit/8c949a6dfffa090d86cc7edf3761f45e328a6404))
+- Fixed a bug where results are not filtered correctly fixes [#117](https://github.com/Amsterdam/Vergunningschecker/issues/117) ([f7a0a62](https://github.com/Amsterdam/Vergunningschecker/commit/f7a0a62692f8dfb8dcb4cf8118daa69b6485e29f))
+- readded toevoeging check ([f924458](https://github.com/Amsterdam/Vergunningschecker/commit/f92445893a28f3263bbf174dc2ce07f20ed7bd65))
+- Removed duplicates from toevoeging ([35b04c9](https://github.com/Amsterdam/Vergunningschecker/commit/35b04c9bb1e3cb75b1bf6b17de513cf4d425b313))
+- Streetnumber is now split ([5b89179](https://github.com/Amsterdam/Vergunningschecker/commit/5b89179edeae918c0a64426808b9542eabde7baf))
+- Validated postcode with space at end ([35053af](https://github.com/Amsterdam/Vergunningschecker/commit/35053af37db8ac8450bc95dfbd25d9ac1bf2d912))
+
+### Refactors
+
+- Added better outlining for the versionrc ([84d1293](https://github.com/Amsterdam/Vergunningschecker/commit/84d1293647509603ebea36ecb2634dc0430f3bfa))
+- Removed old data from changelog ([e1db252](https://github.com/Amsterdam/Vergunningschecker/commit/e1db2521fbab2bc4b589f6f5b64a384f306896d5))
+
+## [0.1.0](https://github.com/Amsterdam/Vergunningschecker/compare/v0.0.8...v0.1.0) (2020-01-09)
+
+### Features
+
+- Added more standard version ([1e9f67e](https://github.com/Amsterdam/Vergunningschecker/commit/1e9f67eb1fc89f8082dd2428c92077df24fd3886))